Fix argc passing in an alternative way by changing nargvstr and nenvstr to be unsigned long instead of int. This preserves argument passing conventions on sparc64, and eliminates hacks on powerpc64.

+++ sys/sys/exec.h
@@ -48,9 +48,9 @@
  */
 struct ps_strings {
 	char	**ps_argvstr;	/* first of 0 or more argument strings */
-	unsigned int ps_nargvstr; /* the number of argument strings */
+	unsigned long ps_nargvstr; /* the number of argument strings */
 	char	**ps_envstr;	/* first of 0 or more environment strings */
-	unsigned int ps_nenvstr; /* the number of environment strings */
+	unsigned long ps_nenvstr; /* the number of environment strings */
 };
 
 /*
